Combine backend process roles in `crates-io`
This keeps the existing role implementations and flattens administrative commands beside the server, background worker, and monitor commands.
