+++ Cargo.toml
@@ -6,7 +6,7 @@ license = "MIT OR Apache-2.0"
 repository = "https://github.com/rust-lang/crates.io"
 description = "Backend of crates.io"
 edition = "2024"
-default-run = "server"
+default-run = "crates-io"
 
 [workspace]
 members = ["crates/*"]
