+++ src/bin/crates-io/admin/analyze_crates.rs
