+++ src/bin/crates-io/admin/render_og_images.rs
