+++ src/bin/crates-io/admin/sync_index.rs
@@ -1,4 +1,4 @@
-use crate::dialoguer;
+use super::dialoguer;
 use anyhow::Result;
 use chrono::{NaiveDate, NaiveTime, TimeZone, Utc};
 use clap::builder::ArgAction;
